requirements-dev.txt
# Development dependencies for Florida Tax Certificate Sale application

# Main requirements
-r requirements.txt

# Testing dependencies
-r requirements-test.txt

# Development server
watchdog==3.0.0
Flask-DebugToolbar==0.13.1

# Linting and code quality
pylint==2.17.5
flake8==6.1.0
black==23.7.0
mypy==1.4.1
bandit==1.7.5
safety==2.3.5

# Documentation
Sphinx==7.1.2
sphinx-rtd-theme==1.3.0
sphinx-autoapi==2.1.1

# Database development tools
Flask-Migrate==4.0.5
Flask-Admin==1.6.1

# API development
flask-swagger-ui==4.11.1
apispec==6.3.0

# Debugging
ipython==8.14.0
ipdb==0.13.13

# Hook dependencies
colorama==0.4.6
pre-commit==3.3.3

# Other development tools
python-dotenv-vault==0.5.2  # For managing multiple .env files
pyinstrument==4.5.1  # Performance profiling 